requirements.txt
api==0.0.7
asgiref==3.2.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==3.0.4
djangorestframework==3.11.0
idna==2.9
jwt==0.6.1
mysqlclient==1.4.6
nose==1.3.7
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
